requirements.txt
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.5
braintree==4.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.10
django-allauth==0.41.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-redis-cache==2.1.1
django-extensions==2.2.9
django-heroku==0.3.1
django-model-utils==4.0.0
django-paypal==1.0.0
django-rq==2.3.2
django-rq-scheduler==1.1.3
django-storages==1.9.1
entrypoints==0.3
idna==2.8
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
oauthlib==3.0.1
psycopg2==2.8.5
gunicorn==20.0.4
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pep8==1.7.1
pickleshare==0.7.5
Pillow==7.1.2
pyyaml==5.3.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
pycodestyle==2.5.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2018.5
# pywin32==227
# pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.21.0
requests-oauthlib==1.2.0
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.2.4
sentry-sdk==0.14.4
dj-database-url==0.5.0
stripe==2.48.0
sweetify==2.0.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
tweepy==3.8.0
urllib3==1.24.2
wcwidth==0.1.9
webencodings==0.5.1
django-heroku==0.3.1
django-mathfilters==1.0.0
widgetsnbextension==3.5.1
boto3==1.13.11
django-storages==1.9.1
pandas==1.0.3
redis>=3
rq>=0.13